+++ babel.config.js
@@ -11,5 +11,7 @@ module.exports = {
         extensions: ['.ios.js', '.android.js', '.js', '.jsx', '.json', '.tsx', '.ts', '.native.js'],
       },
     ],
+    'react-native-reanimated/plugin',
+    'babel-plugin-styled-components',
   ],
 };