+++ src/config/constants.ts
@@ -1,0 +1,6 @@
+import { Dimensions } from 'react-native';
+
+const { width } = Dimensions.get('screen');
+
+export const CARD_WIDTH_FULL = width * 0.85;
+export const IMAGE_IN_LIST = 80;