+++ src/interfaces/author.interface.ts
@@ -1,0 +1,5 @@
+export interface IAuthor {
+  name: string;
+  img: string;
+  description: string;
+}