+++ src/interfaces/dto/posts.dto.ts
@@ -1,0 +1,9 @@
+import { IThrowError } from '~interfaces/dto/shared.dto';
+
+export interface IGetPostsResponse extends IThrowError {
+  posts: any[];
+  meta: {
+    page: number;
+    hasMore: boolean;
+  };
+}