+++ src/interfaces/dto/shared.dto.ts
@@ -1,0 +1,8 @@
+export interface IGenericResponse extends IThrowError {
+  success: boolean;
+  msg?: string;
+}
+
+export interface IThrowError {
+  thrownError?: boolean;
+}