+++ src/interfaces/post.interface.ts
@@ -1,0 +1,7 @@
+import { IAuthor } from '~interfaces/author.interface';
+
+export interface IPost {
+  id: number;
+  author: IAuthor;
+  body: string;
+}