+++ src/interfaces/store.interface.ts
@@ -1,3 +1,7 @@
 export interface SetBooleanPayload {
   value: boolean;
 }
+
+export interface SetNumberPayload {
+  value: number;
+}