+++ src/theme/theme.tsx
@@ -1,6 +1,27 @@
 const theme = {
   colors: {
-    backgroundColor: '#f2f2f6',
+    primary: '#04060A',
+    background: '#f2f2f6',
+    white: '#ffffff',
+  },
+  spacing: {
+    xxs: 2,
+    xs: 4,
+    s: 8,
+    sm: 12,
+    m: 16,
+    l: 24,
+    ml: 32,
+    xl: 40,
+    xxl: 54,
+  },
+  borderRadii: {
+    s: 4,
+    sm: 8,
+    m: 10,
+    ml: 12,
+    l: 25,
+    xl: 75,
   },
 };